Move ZoneUtil to turfgame.util.
