+++ src/test/java/org/joelson/turf/zundin/MonthlyTest.java
@@ -3,7 +3,7 @@ package org.joelson.turf.zundin;
 import org.joelson.turf.statistics.Statistics;
 import org.joelson.turf.statistics.StatisticsInitializer;
 import org.joelson.turf.turfgame.apiv4.Zone;
-import org.joelson.turf.turfgame.apiv4.ZoneUtil;
+import org.joelson.turf.turfgame.util.ZoneUtil;
 import org.joelson.turf.turfgame.apiv4.ZonesTest;
 import org.joelson.turf.util.URLReaderTest;
 import org.junit.jupiter.api.Test;
